util/helpers.py refactor: - remove unused code; - move elsewhere code used once; - improve test coverage.
