+++ decred/tests/integration/dcr/test_rpc.py
@@ -3,9 +3,12 @@ Copyright (c) 2019-2020, the Decred developers
 See LICENSE for details
 """
 
+import configparser
 import os
+import platform
 from urllib.parse import urlunsplit
 
+from appdirs import AppDirs
 from base58 import b58decode
 import pytest
 
@@ -16,17 +19,16 @@ from decred.dcr.nets import mainnet
 from decred.dcr.wire import wire
 from decred.dcr.wire.msgblock import BlockHeader
 from decred.dcr.wire.msgtx import MsgTx, OutPoint, TxIn, TxOut
-from decred.util import helpers
 from decred.util.encode import ByteArray
 
 
 @pytest.fixture
 def config():
-    dcrdCfgDir = helpers.appDataDir("dcrd")
+    dcrdCfgDir = appDataDir("dcrd")
     cfgPath = os.path.join(dcrdCfgDir, "dcrd.conf")
     if not os.path.isfile(cfgPath):
         return None
-    cfg = helpers.readINI(cfgPath, ["rpcuser", "rpcpass", "rpccert"])
+    cfg = readINI(cfgPath, ["rpcuser", "rpcpass", "rpccert"])
     assert "rpcuser" in cfg
     assert "rpcpass" in cfg
     if "rpccert" not in cfg:
@@ -34,6 +36,73 @@ def config():
     if "rpclisten" not in cfg:
         cfg["rpclisten"] = "localhost:9109"
     return cfg
+
+
+def appDataDir(appName):
+    """
+    appDataDir returns an operating system specific directory to be used for
+    storing application data for an application.
+    """
+    if appName == "" or appName == ".":
+        return "."
+
+    # The caller really shouldn't prepend the appName with a period, but
+    # if they do, handle it gracefully by stripping it.
+    appName = appName.lstrip()
+    appNameUpper = appName.capitalize()
+    appNameLower = appName.lower()
+
+    # Get the OS specific home directory.
+    homeDir = os.path.expanduser("~")
+
+    # Fall back to standard HOME environment variable that works
+    # for most POSIX OSes.
+    if homeDir == "":
+        homeDir = os.getenv("HOME")
+
+    opSys = platform.system()
+    if opSys == "Windows":
+        # Windows XP and before didn't have a LOCALAPPDATA, so fallback
+        # to regular APPDATA when LOCALAPPDATA is not set.
+        return AppDirs(appNameUpper, "").user_data_dir
+
+    elif opSys == "Darwin":
+        if homeDir != "":
+            return os.path.join(homeDir, "Library", "Application Support", appNameUpper)
+
+    else:
+        if homeDir != "":
+            return os.path.join(homeDir, "." + appNameLower)
+
+    # Fall back to the current directory if all else fails.
+    return "."
+
+
+def readINI(path, keys):
+    """
+    Attempt to read the specified keys from the INI-formatted configuration
+    file. All sections will be searched. A dict with discovered keys and
+    values will be returned. If a key is not discovered, it will not be
+    present in the result.
+
+    Args:
+        path (str): The path to the INI configuration file.
+        keys (list(str)): Keys to search for.
+
+    Returns:
+        dict: Discovered keys and values.
+    """
+    config = configparser.ConfigParser()
+    # Need to add a section header since configparser doesn't handle sectionless
+    # INI format.
+    with open(path) as f:
+        config.read_string("[tinydecred]\n" + f.read())  # This line does the trick.
+    res = {}
+    for section in config.sections():
+        for k in config[section]:
+            if k in keys:
+                res[k] = config[section][k]
+    return res
 
 
 mainnetAddress = "Dcur2mcGjmENx4DhNqDctW5wJCVyT3Qeqkx"
